Rename and integrate `SoloScoreV2` into existing processors and tests
