+++ osu.Server.Queues.ScoreStatisticsProcessor/Processors/HitStatisticsProcessor.cs
@@ -14,22 +14,22 @@ namespace osu.Server.Queues.ScoreStatisticsProcessor.Processors
     [UsedImplicitly]
     public class HitStatisticsProcessor : IProcessor
     {
-        public void RevertFromUserStats(SoloScore score, UserStats userStats, int previousVersion, MySqlConnection conn, MySqlTransaction transaction)
+        public void RevertFromUserStats(SoloScoreInfo score, UserStats userStats, int previousVersion, MySqlConnection conn, MySqlTransaction transaction)
         {
             if (previousVersion >= 2)
                 adjustStatisticsFromScore(score, userStats, true);
         }
 
-        public void ApplyToUserStats(SoloScore score, UserStats userStats, MySqlConnection conn, MySqlTransaction transaction)
+        public void ApplyToUserStats(SoloScoreInfo score, UserStats userStats, MySqlConnection conn, MySqlTransaction transaction)
         {
             adjustStatisticsFromScore(score, userStats);
         }
 
-        public void ApplyGlobal(SoloScore score, MySqlConnection conn)
+        public void ApplyGlobal(SoloScoreInfo score, MySqlConnection conn)
         {
         }
 
-        private static void adjustStatisticsFromScore(SoloScore score, UserStats userStats, bool revert = false)
+        private static void adjustStatisticsFromScore(SoloScoreInfo score, UserStats userStats, bool revert = false)
         {
             int multiplier = revert ? -1 : 1;
